pyproject.toml
[tool.poetry]
name = "todoister"
version = "0.3.0"
description = ""
authors = ["Gabriel Lopes Rodrigues <gabriellopes@usp.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
todoist-api-python = "^2.0.1"
python-dotenv = "^0.21.0"
click = "^8.1.7"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
ruff = "^0.1.6"

[tool.poetry.scripts]
todoister = "todoister.cli:add_task"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
